Develop

+++ src/controllers/PostController.php
@@ -16,21 +16,22 @@ class PostController extends Controller
     /**
      * Home page controller
      *
-     * @return void
      */
     public function index()
     {
-        $this->render('frontend/homeView.twig');
+        $PostManager = new PostManager($this->db);
+        $posts = $PostManager->getList();
+        $this->render('frontend/homeView.twig', array(
+            'posts' => $posts
+        ));
     }
 
     /**
      * Blog page controller
      *
-     * @return void
      */
     public function blog()
     {
-        var_dump($_GET);
         $PostManager = new PostManager($this->db);
         $posts = $PostManager->getList();
         $this->render(
@@ -43,7 +44,6 @@ class PostController extends Controller
     /**
      * Read a single post controller
      *
-     * @return void
      */
     public function single()
     {
@@ -62,14 +62,14 @@ class PostController extends Controller
                 )
             );
         } else {
-            echo 'Cet article n\'existe pas !';
+            $errorController = new ErrorController();
+            $errorController->error404();
         }
     }
 
     /**
      * Home admin page controller
      *
-     * @return void
      */
     public function adminIndex()
     {
@@ -93,7 +93,6 @@ class PostController extends Controller
     /**
      * Add post controller
      *
-     * @return void
      */
     public function create()
     {
@@ -105,16 +104,16 @@ class PostController extends Controller
                 $_POST, ['username', 'title', 'teaser', 'content']
             ) 
             ) {
-                if ($this->tokenValidate("http://localhost/Projet5/admin/ajouter", 30)) {
+                if ($this->tokenValidate("http://localhost/Projet5/admin/ajouter", 300)) {
                     $imagePath = Image::getImage('post');
                     $post = new Post(
                         [
-                        'author' => htmlspecialchars($_POST['username']),
-                        'title' => htmlspecialchars($_POST['title']),
-                        'teaser' => htmlspecialchars($_POST['teaser']),
-                        'content' => htmlspecialchars($_POST['content']),
+                        'author' => htmlspecialchars($_POST['username'], ENT_NOQUOTES),
+                        'title' => htmlspecialchars($_POST['title'], ENT_NOQUOTES),
+                        'teaser' => htmlspecialchars($_POST['teaser'], ENT_NOQUOTES),
+                        'content' => htmlspecialchars($_POST['content'], ENT_NOQUOTES),
                         'imagePath' => $imagePath,
-                        'slug' => htmlspecialchars($_POST['title']),
+                        'slug' => htmlspecialchars($_POST['title'], ENT_NOQUOTES),
                         'validComment' => 0,
                         'newComment' => 0
                         ]
@@ -146,7 +145,6 @@ class PostController extends Controller
     /**
      * Read a post controller in administration
      *
-     * @return void
      */
     public function read()
     {
@@ -168,7 +166,8 @@ class PostController extends Controller
                     )
                 );
             } else {
-                echo 'Cet article n\'existe pas !';
+                $errorController = new ErrorController();
+                $errorController->error404();
             }
             
         } else {
@@ -193,7 +192,6 @@ class PostController extends Controller
     /**
      * Update a post controller in administration
      *
-     * @return void
      */
     public function update()
     {
@@ -214,17 +212,17 @@ class PostController extends Controller
                     $_POST, ['username', 'title', 'teaser', 'content'] 
                 )
                 ) {
-                    if ($this->tokenValidate("http://localhost/Projet5/admin/modifier/$slug-$id", 30)) {
+                    if ($this->tokenValidate("http://localhost/Projet5/admin/modifier/$slug-$id", 300)) {
                         $imagePath = Image::getImage('post');
                         $post = new Post(
                             [
                             'id' => $id,
-                            'author' => $_POST['username'],
-                            'title' => $_POST['title'],
-                            'teaser' => $_POST['teaser'],
-                            'content' => $_POST['content'],
+                            'author' => htmlspecialchars($_POST['username'], ENT_NOQUOTES),
+                            'title' => htmlspecialchars($_POST['title'], ENT_NOQUOTES),
+                            'teaser' => htmlspecialchars($_POST['teaser'], ENT_NOQUOTES),
+                            'content' => htmlspecialchars($_POST['content'], ENT_NOQUOTES),
                             'imagePath' => $imagePath,
-                            'slug' => $_POST['title'],
+                            'slug' => htmlspecialchars($_POST['title'], ENT_NOQUOTES),
                             'validComment' => $countValid,
                             'newComment' => $countNew
                             ]
@@ -262,16 +260,21 @@ class PostController extends Controller
     /**
      * Delete a post controller in administration
      *
-     * @return void
      */
     public function delete()
     {
         $PostManager = new PostManager($this->db);
 
+        $id = (int)substr(strrchr($_SERVER['REQUEST_URI'], '-'), 1);
+
         if ($this->sessionExist('user', 'ADMIN')) {
-            $id = (int)substr(strrchr($_SERVER['REQUEST_URI'], '-'), 1);
-            $PostManager->delete($id);
-            header('location:../../admin');
+            if ($this->tokenValidate("http://localhost/Projet5/admin", 300)) {
+                $PostManager->delete($id);
+                header('location:http://localhost/Projet5/admin');
+            } else {
+                session_unset();
+                header('location: http://localhost/Projet5');
+            } 
         } else {
             header('location: http://localhost/Projet5');
         }
